\name{deeplearning-class}
\docType{class}
\alias{deeplearning-class}
\alias{da-class}
\alias{sda-class}
\alias{Rcpp_Sda-class}
\alias{Rcpp_dA-class}
\alias{class:deeplearning}
\alias{class:sda}
\alias{class:Rcpp_Sda}
\alias{class:da}
\alias{class:Rcpp_dA}

\title{Class \code{deeplearning} and sub-classes}

\author{
  Qiang Kou
}

\keyword{classes}
