\docType{data}
\name{datasets}
\alias{train_X}
\alias{train_Y}
\alias{test_X}
\alias{hidden}
\title{Testing dataset for deep learning methods}
\usage{
data(test)
}
\description{
A testing dataset for deep learning methods:
}
\details{
\itemize{
  \item \code{train_X}. The training dataset. 
  \item \code{train_y}. The labels for training dataset.
  \item \code{test_X}. The testing dataset.
  \item \code{hidden}. The number of hidden representation in each layer.
}
}
\keyword{datasets}
