\name{RcppDL-package}
\alias{RcppDL-package}
\alias{RcppDL}
\docType{package}
\title{
Deep Learning Methods via Rcpp
}
\description{
This package is based on the C++ code from Yusuke Sugomori,
which implements basic machine learning methods with 
many layers (deep learning), including dA (Denoising Autoencoder), 
SdA (Stacked Denoising Autoencoder), RBM (Restricted Boltzmann machine) and 
DBN (Deep Belief Nets).
}
\details{
\tabular{ll}{
Package: \tab RcppDL\cr
Type: \tab Package\cr
Version: \tab 0.0.5\cr
Date: \tab 2014-12-12\cr
License: \tab GPL-3 \cr
}

}
\author{
Qiang Kou, Yusuke Sugomori

Maintainer: Qiang Kou <qkou@umail.iu.edu>
}
\references{
Y. Bengio, P. Lamblin, D. Popovici, H. Larochelle: Greedy Layer-Wise
Training of Deep Networks, Advances in Neural Information Processing
Systems 19, 2007

P. Vincent, H. Larochelle, Y. Bengio, P.A. Manzagol: Extracting and
Composing Robust Features with Denoising Autoencoders, ICML' 08, 1096-1103, 2008
}

\keyword{ package }
